adding "lato" .ttf font files
This commmit adds a tarball with all the .ttf files in the lato directory, as included on Debian 12. The tarball was created with the following command:

  tar -cjvf lato.tar.bz2 /usr/share/fonts/truetype/lato/*

Note that the checksum of the font file that we're using for the logo (Lato-Bold.ttf) is distinct from both the one available from the font's official website and the one available from Google Fonts. But we're using the one in Debian in inkscape, so that's the one we're adding to this repo.

 * https://www.latofonts.com/
 * https://fonts.google.com/specimen/Lato

		user@disp7670:~/Downloads$ sha256sum Lato/Lato-Bold.ttf
		d7f0b7f2570f2f28b504da1181b4d71b1420b10be2c4fd690927f1c8ee3b19c3  Lato/Lato-Bold.ttf
		user@disp7670:~/Downloads$

		user@disp7670:~/Downloads$ sha256sum ./Lato2OFL/Lato2OFL/Lato-Bold.ttf
		bf1b8130069b44b9148eeece35e5423bedac49777ba746615b826b8276574a7b  ./Lato2OFL/Lato2OFL/Lato-Bold.ttf
		user@disp7670:~/Downloads$

		user@disp7670:~/Downloads$ sha256sum /usr/share/fonts/truetype/lato/Lato-Bold.ttf
		e0b1c4de011de8dd077cf21f97919fc31158eb290ee1176cbae06d500a6304bc  /usr/share/fonts/truetype/lato/Lato-Bold.ttf
		user@disp7670:~/Downloads$